+++ rundeck/resource_job.go
@@ -125,10 +125,42 @@ func resourceRundeckJob() *schema.Resource {
 				Type:     schema.TypeBool,
 				Optional: true,
 			},
-
 			"timeout": {
 				Type:     schema.TypeString,
 				Optional: true,
+			},
+			"orchestrator": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"type": {
+							Type:        schema.TypeString,
+							Required:    true,
+							Description: "Option of `subset`, `rankTiered`, `maxPercentage`, `orchestrator-highest-lowest-attribute`",
+						},
+						"count": {
+							Type:        schema.TypeInt,
+							Optional:    true,
+							Description: "Value for the subset orchestrator",
+						},
+						"percent": {
+							Type:        schema.TypeInt,
+							Optional:    true,
+							Description: "Value for the maxPercentage orchestrator",
+						},
+						"attribute": {
+							Type:        schema.TypeString,
+							Optional:    true,
+							Description: "The Node Attribute that shoud be used to rank nodes in High/Low Orchestrator.",
+						},
+						"sort": {
+							Type:        schema.TypeString,
+							Optional:    true,
+							Description: "Option of `highest` or `lowest` for High/Low Orchestrator",
+						},
+					},
+				},
 			},
 
 			"schedule": {
@@ -278,6 +310,10 @@ func resourceRundeckJob() *schema.Resource {
 						},
 						"storage_path": {
 							Type:     schema.TypeString,
+							Optional: true,
+						},
+						"hidden": {
+							Type:     schema.TypeBool,
 							Optional: true,
 						},
 					},
@@ -627,6 +663,49 @@ func jobFromResourceData(d *schema.ResourceData) (*JobDetail, error) {
 		job.Dispatch.SuccessOnEmptyNodeFilter = successOnEmpty.(bool)
 	}
 
+	orchList := d.Get("orchestrator").([]interface{})
+	if len(orchList) > 1 {
+		return nil, fmt.Errorf("rundeck command may have no more than one orchestrator")
+	}
+	for _, orch := range orchList {
+		orchMap := orch.(map[string]interface{})
+		job.Orchestrator = &JobOrchestrator{
+			Type:   orchMap["type"].(string),
+			Config: JobOrchestratorConfig{},
+		}
+		orchType := orchMap["type"].(string)
+		if orchType == "orchestrator-highest-lowest-attribute" {
+			orchAttr := orchMap["attribute"]
+			if orchAttr != nil {
+				job.Orchestrator.Config.Attribute = orchAttr.(string)
+			} else {
+				return nil, fmt.Errorf("high Low Orchestrator must include an attribute to sort against")
+			}
+			orchSort := orchMap["sort"]
+			if orchSort != nil {
+				job.Orchestrator.Config.Sort = orchSort.(string)
+			} else {
+				return nil, fmt.Errorf("high low orchestrator must include sort direction of `high` or `low`")
+			}
+		}
+		if orchType == "subset" {
+			orchCount := orchMap["count"]
+			if orchCount != nil {
+				job.Orchestrator.Config.Count = orchCount.(int)
+			} else {
+				return nil, fmt.Errorf("subset Orchestrator requires count setting")
+			}
+		}
+		if orchType == "maxPercentage" {
+			orchPct := orchMap["percent"]
+			if orchPct != nil {
+				job.Orchestrator.Config.Percent = orchPct.(int)
+			} else {
+				return nil, fmt.Errorf("max Percentage Orchestrator requires a percent integer configuration")
+			}
+		}
+	}
+
 	sequence := &JobCommandSequence{
 		ContinueOnError:  d.Get("continue_on_error").(bool),
 		OrderingStrategy: d.Get("command_ordering_strategy").(string),
@@ -686,12 +765,13 @@ func jobFromResourceData(d *schema.ResourceData) (*JobDetail, error) {
 				ObscureInput:            optionMap["obscure_input"].(bool),
 				ValueIsExposedToScripts: optionMap["exposed_to_scripts"].(bool),
 				StoragePath:             optionMap["storage_path"].(string),
+				Hidden:                  optionMap["hidden"].(bool),
 			}
-			if option.StoragePath != "" && option.ObscureInput == false {
-				return nil, fmt.Errorf("Argument \"obscure_input\" must be set to `true` when \"storage_path\" is not empty.")
+			if option.StoragePath != "" && !option.ObscureInput {
+				return nil, fmt.Errorf("argument \"obscure_input\" must be set to `true` when \"storage_path\" is not empty")
 			}
-			if option.ValueIsExposedToScripts && option.ObscureInput == false {
-				return nil, fmt.Errorf("Argument \"obscure_input\" must be set to `true` when \"exposed_to_scripts\" is set to true.")
+			if option.ValueIsExposedToScripts && !option.ObscureInput {
+				return nil, fmt.Errorf("argument \"obscure_input\" must be set to `true` when \"exposed_to_scripts\" is set to true")
 			}
 
 			for _, iv := range optionMap["value_choices"].([]interface{}) {
@@ -924,6 +1004,7 @@ func jobToResourceData(job *JobDetail, d *schema.ResourceData) error {
 				"obscure_input":             option.ObscureInput,
 				"exposed_to_scripts":        option.ValueIsExposedToScripts,
 				"storage_path":              option.StoragePath,
+				"hidden":                    option.Hidden,
 			}
 			optionConfigsI = append(optionConfigsI, optionConfigI)
 		}